Api.cs implements CreateSecureHash and SHA256 as static methods
