--- Common/Util/ApiExtensions.cs
@@ -1,53 +1,0 @@
-﻿/*
- * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
- * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
-*/
-
-using System.Security.Cryptography;
-using System.Text;
-
-namespace QuantConnect.Util
-{
-    public class ApiExtensions
-    {
-        /// <summary>
-        /// Generate a secure hash for the authorization headers.
-        /// </summary>
-        /// <returns>Time based hash of user token and timestamp.</returns>
-        public static string CreateSecureHash(int timestamp, string token)
-        {
-            // Create a new hash using current UTC timestamp.
-            // Hash must be generated fresh each time.
-            var data = string.Format("{0}:{1}", token, timestamp);
-            return SHA256(data);
-        }
-
-        /// <summary>
-        /// Encrypt the token:time data to make our API hash.
-        /// </summary>
-        /// <param name="data">Data to be hashed by SHA256</param>
-        /// <returns>Hashed string.</returns>
-        public static string SHA256(string data)
-        {
-            var crypt = new SHA256Managed();
-            var hash = new StringBuilder();
-            var crypto = crypt.ComputeHash(Encoding.UTF8.GetBytes(data), 0, Encoding.UTF8.GetByteCount(data));
-            foreach (var theByte in crypto)
-            {
-                hash.Append(theByte.ToString("x2"));
-            }
-            return hash.ToString();
-        }
-    }
-}
